Create main.py
This source code applies the SLIC algorithm in a given input image with 3 channels. The difference between our algorithm and the original one (by Achanta et al. 2010) is that they use CIELAB space, whereas we apply RGB original space.
To use this algorithm you need to create a folder called 'animation' in which the partial results will be stored. Our algorithm is not optimized for high performance, instead it is used only to run once to create a beautiful animation (series of png files), which can be converted to a swf file, for example, using png2swg.